+++ src/plugins/scope.ts
@@ -4,7 +4,7 @@ import { ConfigurationOptions } from '../entities/configuration';
 import { Option } from '../utils/types';
 import Key from '../utils/key';
 
-export interface Configuration extends ConfigurationOptions {
+export interface ScopeConfiguration extends ConfigurationOptions {
     scopes: {
         only: boolean | undefined;
         list: Option;
@@ -15,7 +15,7 @@ export default class ScopePlugin extends CommitPlugin {
     private scopes: Map<string, string> = new Map();
     private onlyConfigured: boolean = false;
 
-    public async init(config: Configuration): Promise<void> {
+    public async init(config: ScopeConfiguration): Promise<void> {
         const { scopes } = config;
 
         if (typeof scopes === 'object') {