+++ src/plugins/section.ts
@@ -5,14 +5,14 @@ import { Option, OptionValue } from '../utils/types';
 import { Section, Position } from '../entities/section';
 import Key from '../utils/key';
 
-export interface Configuration extends ConfigurationOptions {
+export interface SectionConfiguration extends ConfigurationOptions {
     sections: Option;
 }
 
 export default class SectionPlugin extends CommitPlugin {
     private blocks: Map<string, Section> = new Map();
 
-    public async init(config: Configuration): Promise<void> {
+    public async init(config: SectionConfiguration): Promise<void> {
         const { sections } = config;
 
         if (Array.isArray(sections)) {