+++ test/plugins/attention.test.ts
@@ -1,29 +1,37 @@
 import { Task } from 'tasktree-cli/lib/task';
 import { MockState } from '../__mocks__/entities/state.mock';
 import { Configuration } from '../../src/entities/configuration';
+import { Dependency, DependencyType } from '../../src/entities/package/dependency';
+import AttentionPlugin, { AttentionConfiguration } from '../../src/plugins/attention';
 import { Level } from '../../src/utils/enums';
-import { Message } from '../../src/entities/message';
-import AttentionPlugin, { Configuration as AttentionConfiguration } from '../../src/plugins/attention';
 
 describe('AttentionPlugin', (): void => {
-    const config = new Configuration();
-    const context = new MockState();
-    const plugin = new AttentionPlugin(context);
-    const task = new Task('test task');
+    let config: Configuration;
+    let context: MockState;
+    let plugin: AttentionPlugin;
+    let task: Task;
 
-    it('Default', (done): void => {
-        context.setLicense('MIT', undefined);
+    beforeEach((done): void => {
+        config = new Configuration();
+        context = new MockState();
+        plugin = new AttentionPlugin(context);
+        task = new Task('test task');
+
         config.load(task).then((): void => {
             plugin.init(config.getOptions() as AttentionConfiguration).then((): void => {
-                expect(context.getSections().length).toBe(1);
-                expect(context.findSection('Important Changes')).toBeDefined();
-
                 done();
             });
         });
     });
 
+    it('Default', (): void => {
+        expect(context.getSections().length).toBe(1);
+        expect(context.findSection('Important Changes')).toBeDefined();
+    });
+
     it('Changed license', (done): void => {
+        context.setLicense('MIT', undefined);
+
         const section = context.findSection('Important Changes');
         const license = context.getLicense();
 
@@ -32,20 +40,45 @@ describe('AttentionPlugin', (): void => {
 
         if (section && license) {
             plugin.modify(task).then((): void => {
-                const getData = (l: Message[]): [string, number][] => l.map((m): [string, number] => [m.text, m.level]);
-                const messages = new Map(
-                    section.getSections().map((s): [string, Message[]] => [s.title, s.getMessages()])
-                );
-                const licenseMessages = messages.get('License');
+                const [subsection] = section.getSections();
 
-                expect(messages.size).toBe(1);
-                expect(licenseMessages).toBeDefined();
+                expect(subsection).toBeDefined();
+                expect(subsection.title).toBe('License');
+                expect(
+                    subsection.getMessages().map((message): [Level, string] => [message.level, message.text])
+                ).toMatchSnapshot();
 
-                if (licenseMessages) {
-                    expect(getData(licenseMessages)).toStrictEqual([
-                        [['Source code now under `MIT` license.'].join(''), Level.Major],
-                    ]);
+                done();
+            });
+        }
+    });
+
+    it('Changed dependencies', (done): void => {
+        context.setDependencies(
+            new Dependency(
+                DependencyType.Dependencies,
+                {
+                    test: '^1.1.0',
+                },
+                {
+                    test: '^1.0.0',
                 }
+            )
+        );
+
+        const section = context.findSection('Important Changes');
+
+        expect(section).toBeDefined();
+
+        if (section) {
+            plugin.modify(task).then((): void => {
+                const [subsection] = section.getSections();
+
+                expect(subsection).toBeDefined();
+                expect(subsection.title).toBe('Dependencies');
+                expect(
+                    subsection.getMessages().map((message): [Level, string] => [message.level, message.text])
+                ).toMatchSnapshot();
 
                 done();
             });