+++ test/plugins/section.test.ts
@@ -2,7 +2,7 @@ import { Task } from 'tasktree-cli/lib/task';
 import { MockState } from '../__mocks__/entities/state.mock';
 import { Configuration } from '../../src/entities/configuration';
 import { Commit } from '../../src/entities/commit';
-import SectionPlugin, { Configuration as SectionConfiguration } from '../../src/plugins/section';
+import SectionPlugin, { SectionConfiguration } from '../../src/plugins/section';
 
 describe('SectionPlugin', (): void => {
     const context = new MockState();
@@ -13,9 +13,10 @@ describe('SectionPlugin', (): void => {
     it('Default', (done): void => {
         config.load(task).then((): void => {
             plugin.init(config.getOptions() as SectionConfiguration).then((): void => {
-                expect(context.getSections().length).toBe(6);
-                expect(context.findSection('Bug Fixes')).toBeDefined();
+                expect(context.getSections().length).toBe(7);
                 expect(context.findSection('Features')).toBeDefined();
+                expect(context.findSection('Improvements')).toBeDefined();
+                expect(context.findSection('Bug Fixes')).toBeDefined();
                 expect(context.findSection('Internal changes')).toBeDefined();
                 expect(context.findSection('Performance Improvements')).toBeDefined();
                 expect(context.findSection('Code Refactoring')).toBeDefined();